requirements.txt
torch==2.6.0
torchvision==0.21.0
pillow==11.1.0